Moved relevant files to ActionItemModal so name doesn't conflict with master
